Bugfix: explicitly define dtype
if all the entries were integers, an error would be thrown on attempting
to fill with np.inf (a float)

+++ dacapo/run.py
@@ -183,7 +183,8 @@ class Run:
                                     ]
                                 )
                                 for parameter_scores in self.validation_scores.scores
-                            ]
+                            ],
+                            dtype=np.float32,
                         )
 
                         # replace nan
